pyproject.toml
[build-system]
requires = ["pdm-backend >= 2.4.0"]
build-backend = "pdm.backend"
[project]
name = "goodass"
version = "0.0.1"
authors = [
  { name="Nicholas Ciardiello", email="nicholasciardiello2000@gmail.com" },
]
description = "Global Oversight Of Deployed Authorized SSH Settings"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "prettytable==3.17.0",
    "PyYAML==6.0.3",
    "paramiko==4.0.0",
    "platformdirs==4.5.0",
]
license = "Apache-2.0"
license-files = ["LICEN[CS]E*"]



[project.scripts]
goodass = "goodass.cli:main"

[project.urls]
Homepage = "https://github.com/Nidhil-stack/G.O.O.D.A.S.S."
Issues = "https://github.com/Nidhil-stack/G.O.O.D.A.S.S./issues"
